requirements.txt
pypdf
pathlib
sentence_transformers
faiss-cpu
numpy
langchain
langchain_core
langchain_community
langchain_text_splitters
langchain_huggingface
langchain_openai